Iscsi storage take2 (#35)
iSCSI support for XenServer and VMware:

iSCSI storage server requires targetcli.
For XenServer iSCSI SR's will be configured by CloudStack.
For VMware iSCSI LUNs are formatted and added to the cluster before zone create script runs.
Please note - iSCSI support for KVM has deliberately not been included. This requires CLVM or similar clustered file system, which during all my testing has proven to be unstable and not fit for purpose. Support for this may be included in Trillian in the future.